Change namespaces and class names
This is a non-backward-compatible change. I have taken the fork in a
new direction. For new installations this will work just fine. But, it
will not work for previous installations of yajra's code. This patch
does the following:

  - Changes the namespace root from yajra to alfmel
  - Changes the namespace of the code from PDO to OCI8
  - Renames yajra\Pdo\Oci8 class to alfmel\OCI8\PDO
  - Renames yajra\Pdo\Oci8\Statement to alfmel\OCI8\PDOStatement
  - Removes the SqlException; all code now throws PDOException
  - Removes the src directory and moves the alfmel directory up one level
  - Change sthe README.md file to show I'm the maintainer of this fork

I think the class name changes make it more consistent with the original
structure of the built-in PDO classes. It also makes the directory
structure much cleaner and actually removes the need for use statements.